requirements.txt
certifi==2023.7.22
chardet==3.0.4
colorama==0.4.6
configparser==6.0.0
crayons==0.4.0
idna==2.10
python-dotenv==1.0.0
requests==2.24.0
selenium==3.141.0
urllib3==1.25.11
webdriver-manager==3.4.2
